+++ api/resources/notification.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 # standard python imports
 from flask_restful import Resource, reqparse
-#from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, current_user
 from models.notification import NotificationModel
 #from app.util.logz import create_logger
 
@@ -12,27 +12,27 @@ class Notification(Resource):
     parse.add_argument('message', type=str, required=True, help="This field cannot be left blank!")
     parse.add_argument('read', type=bool, required=False, help="This field cannot be left blank!")
 
-
     def __init__(self):
         #self.logger = create_logger(__name__)
         pass
 
+    @jwt_required
     def get(self, id):
         notification = NotificationModel.find_by_id(id)
         #self.logger.info("Notification get: {}".format(notification))
-        if notification:
+        if notification and (notification.json()['institution'] == current_user.json()['institution']):
             return notification.json()
         return {'message': 'Notification not found'}, 404
 
-    #@jwt_required
+    @jwt_required
     def post(self, id):
         data = self.parse.parse_args()
-        notification = NotificationModel.find_by_id(id)
+        notification = NotificationModel.find_by_id(data['id'])
 
         if notification:
             return {'message': "An notification with id '{}' already exists.".format(id)}, 400
 
-        notification = NotificationModel(**data)
+        notification = NotificationModel(**data, institution_id=current_user.json()['institution'])
 
         try:
             notification.save_to_db()
@@ -41,29 +41,34 @@ class Notification(Resource):
 
         return notification.json(), 201
 
-    #@jwt_required
+    @jwt_required
     def put(self, id):
         data = self.parse.parse_args()
-        notification = NotificationModel.find_by_id(id)
+        notification = NotificationModel.find_by_id(data['id'])
 
-        if notification:
-            notification.type = data['type']
-            notification.service_id = data['service_id']
+        if notification.json()['institution'] == current_user.json()['institution']:
+            if notification:
+                notification.medical_doctor_id = data['medical_doctor_id']
+                notification.read = data['read']
+                notification.message = data['message']
+            else:
+                notification = NotificationModel(**data, institution_id=current_user.json()['institution'])
+
+            notification.save_to_db()
+
+            return notification.json()
         else:
-            notification = NotificationModel(**data)
+            return {'message': 'access denied'}, 401
 
-        notification.save_to_db()
-
-        return notification.json()
-
-    #@jwt_required
+    @jwt_required
     def delete(self, id):
         notification = NotificationModel.find_by_id(id)
-        if notification:
+        if notification and (notification.json()['institution'] == current_user.json()['institution']):
             notification.delete_from_db()
             return {'message': 'Notification deleted'}
         return {'message': 'Notification not found'}, 404
 
 class NotificationList(Resource):
+    @jwt_required
     def get(self):
-        return {'notifications': [notification.json() for notification in Notification.query.all()]}
+        return {'notifications': [notification.json() for notification in Notification.query.filter_by(institution_id=current_user.json()['institution']).all()]}
